RMX3031: fod_extension: fix build after LineageOS/android_frameworks_native@6f55c22fd40b56dc1080905c1161efb4a3a7e756 MIME-Version: 1.0 Content-Type: text/plain; charset=UTF-8 Content-Transfer-Encoding: 8bit
-> changed the data types ¯\_(ツ)_/¯

+++ fod/FodExtension.cpp
@@ -20,6 +20,6 @@ uint32_t getFodZOrder(uint32_t z, bool touched) {
     return touched ? 1090519091 : 1090519089;
 }
 
-uint32_t getFodUsageBits(uint32_t usageBits, bool) {
+uint64_t getFodUsageBits(uint64_t usageBits, bool) {
     return usageBits;
 }
